v0.01 - Refactoring

+++ repository/src/main/java/br/com/apps/repository/repository/user/UserInterfaces.kt
@@ -13,7 +13,7 @@ interface UserReadInterface {
 
     suspend fun updateRequestNumber(uid: String)
 
-    suspend fun getUserRequestNumber(uid: String): Int
+    suspend fun getUserRequestNumber(uid: String): Long
 
 }
